+++ scylla/src/lib.rs
@@ -258,7 +258,6 @@ pub mod cloud;
 pub mod cluster;
 pub mod connection;
 pub mod execution;
-pub mod history;
 pub mod routing;
 pub mod session;
 pub mod statement;
