+++ scylla/src/transport/iterator.rs
@@ -33,13 +33,13 @@ use crate::session::RequestSpan;
 use crate::cluster::ClusterData;
 use crate::cluster::NodeRef;
 use crate::connection::{Connection, NonErrorQueryResponse, QueryResponse};
+use crate::execution::history::{self, HistoryListener};
 use crate::execution::load_balancing::{self, RoutingInfo};
 use crate::execution::retries::{QueryInfo, RetryDecision, RetrySession};
 use crate::frame::response::{
     result,
     result::{ColumnSpec, Row},
 };
-use crate::history::{self, HistoryListener};
 use crate::statement::{prepared_statement::PreparedStatement, query::Query};
 use crate::transport::errors::ProtocolError;
 use crate::transport::metrics::Metrics;
